+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/KeywordFile.java
@@ -29,6 +29,4 @@ public interface KeywordFile {
    PsiFile getPsiFile();
 
    boolean isDifferentNamespace();
-
-   boolean isValid();
 }